requirements.txt
altair
argon2-cffi
argon2-cffi-bindings
asttokens
attrs
backcall
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
cycler
datetime
debugpy
decorator
defusedxml
entrypoints
et-xmlfile
executing
fastjsonschema
fonttools
gitdb
GitPython
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kiwisolver
MarkupSafe
matplotlib
matplotlib.pyplot
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
openpyxl
packaging
pandas
pandocfilters
parso
pexpect
pickle
pickleshare
Pillow
plotly.express
prometheus-client
prompt-toolkit
protobuf==3.20.*
psutil
ptyprocess
pure-eval
pyarrow
pycparser
pydeck
Pygments
Pympler
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
pyzmq
requests
semver
Send2Trash
six
smmap
soupsieve
stack-data
streamlit
streamlit_extras
tabulate
terminado
tinycss2
toml
toolz
tornado
traitlets
tzdata
tzlocal
urllib3
validators
watchdog
wcwidth
webencodings
widgetsnbextension
zipp
